+++ app/concerns/pages_core/taggable.rb
@@ -13,11 +13,6 @@ module PagesCore
       def tagged_with(*tags)
         all.includes(:tags).where('tags.name IN (?)', Tag.parse(tags)).references(:tags)
       end
-
-      def find_tagged_with(tags)
-        ActiveSupport::Deprecation.warn "find_tagged_with is deprecated, use the tagged_with scope."
-        self.tagged_with(tags)
-      end
     end
 
     def serialized_tags